fix(tui): chat: adjust padding for message list
The gap between the message list and the editor was too large. 3 lines
of padding can be too much. 1 line gets added after each message list
item as a "gap", there's the padding at the bottom of the message list,
and there's the editor top padding.

In this commit, we removed the bottom padding of the message list,
keeping both the gap that gets added after each message list item and
the padding at the top of the editor.